fix code smells + refactoring
