+++ dataland-e2etests/src/test/kotlin/org/dataland/e2etests/utils/testDataProvivders/FrameworkTestDataProvider.kt
@@ -1,10 +1,7 @@
-package org.dataland.e2etests.utils
+package org.dataland.e2etests.utils.testDataProvivders
 
-import com.squareup.moshi.FromJson
-import com.squareup.moshi.Json
 import com.squareup.moshi.JsonAdapter
 import com.squareup.moshi.Moshi
-import com.squareup.moshi.ToJson
 import com.squareup.moshi.Types
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import org.dataland.datalandbackend.openApiClient.model.CompanyInformation
@@ -16,25 +13,7 @@ import org.dataland.datalandbackend.openApiClient.model.PathwaysToParisData
 import org.dataland.datalandbackend.openApiClient.model.SfdrData
 import org.dataland.datalandbackend.openApiClient.model.VsmeData
 import java.io.File
-import java.math.BigDecimal
-import java.time.LocalDate
 import java.util.UUID
-
-object BigDecimalAdapter {
-    @FromJson
-    fun fromJson(string: String) = BigDecimal(string)
-
-    @ToJson
-    fun toJson(value: BigDecimal) = value.toString()
-}
-
-object LocalDateAdapter {
-    @FromJson
-    fun fromJson(string: String) = LocalDate.parse(string)
-
-    @ToJson
-    fun toJson(value: LocalDate) = value.toString()
-}
 
 class FrameworkTestDataProvider<T> (private val clazz: Class<T>) {
 
@@ -145,7 +124,3 @@ class FrameworkTestDataProvider<T> (private val clazz: Class<T>) {
             .map { it.t }
     }
 }
-data class CompanyInformationWithT<T>(
-    @Json var companyInformation: CompanyInformation,
-    @Json var t: T,
-)
